requirements.txt
torch~=1.6.0
torchvision~=0.7.0
numpy~=1.18.1
yaml~=0.1.7
pyyaml~=5.3
tqdm~=4.42.1
pillow~=7.0.0
albumentations~=0.4.5
tifffile~=2020.10.1
loguru~=0.5.3
tensorboardx~=2.1
opencv-python~=4.2.0.34
matplotlib~=3.1.3